+++ AlienFight/Model/Objects/GameObject.cs
@@ -4,8 +4,8 @@
     {
         public float X { get; set; }
         public float Y { get; set; }
-        public int SizeX { get; set; }
-        public int SizeY { get; set; }
+        public float SizeX { get; set; }
+        public float SizeY { get; set; }
         public int State { get; set; }
         public bool FlippedY { get; set; }
     }
